feat(indexer): clickhouse support for bundles and transactions (#10)
This PR introduces an optional Clickhouse indexer configurable from CLI.
It tracks full bundles received from the user endpoints and writes them
to the database.

The goal of this is to being able to replicate real flow in a testing
environment, as such it provides also testing code from converting
clickhouse data to `RawBundle`s.