feat(simulation): support for PCAP files -- monitor bandwidth usage (#31)
Packet Capture (PCAP) is a format for recording network traffic on
files. It is used by tools by `tcpdump`,`tshark` and other APIs.

Given that shadow simulation doesn't rely on the OS and intercept all
syscalls for its emulation, using advanced tools like eBPF for
inspecting sockets or packets is not possible. However, shadow exposes a
feature to save network traffic generated during the simulation to PCAP
files, so we can use that, and this PR adds support for it.