feat: better graceful shutdown from `reth_tasks` (#35)
As suggested in #28.

This PR takes code from `reth_tasks` and from `reth_cli` and adapts it
with very small changes to adapt it to the overflow proxy. I think the
way shutdown is handled in such crates is very well done, and after
reading and understanding it I couldn't think of a better solution in a
reasonable manner, so I decided to adapt it.