feat: better tracing (#128)
Closes #127. It introduces spans for the basic units of
work of the proxy, most notably user and system handlers. This paves the
way for a possible future support of tracing log collectors.

The short id allows group logs together, and we add relevant fields into
the span so we're able to track each order behaviour inside the system
and see which code paths are taken.

Moreover, all relevant logs inside the codebase will report the order
hash. In this way, given a log file, we can query the entire lifetime of
an order by `grep <hash> <log_file>` or `grep <id> <log_file>`