perf: configure Tokio, threadpool for compute (#141)
## Description

The main goal of this PR is to ensure that blocking, compute intensive
tasks (which are the majority of what `spawn_with_priority` is currently
used for) are spawned on dedicated threads that do not interfere with
the tokio runtime. It also explicitly configures Tokio to some values
that we can play around with, and adds metrics to priority task
execution.

After playing around with Tokio's
[`spawn_blocking`](https://docs.rs/tokio/latest/tokio/task/fn.spawn_blocking.html),
we didn't really see a reduction in tail latencies, and a higher cost of
number of threads spawned (30 threads at some point on a 4-core
machine). They also mention that it's probably not the right fit for our
usage:

> If your code is CPU-bound and you wish to limit the number of threads
used to run it, you should use a separate thread pool dedicated to CPU
bound tasks. For example, you could consider using the
[rayon](https://docs.rs/rayon) library for CPU-bound tasks. It is also
possible to create an extra Tokio runtime dedicated to CPU-bound tasks,
but if you do this, you should be careful that the extra runtime runs
only CPU-bound tasks, as IO-bound tasks on that runtime will behave
poorly.
> Hint: If using rayon, you can use a
[oneshot](https://docs.rs/tokio/latest/tokio/sync/oneshot/index.html)
channel to send the result back to Tokio when the rayon task finishes.


https://docs.rs/tokio/latest/tokio/index.html#cpu-bound-tasks-and-blocking-code

We settled on a Rayon threadpool with 2 worker threads, and have yet to
see those threads hit their limit on the staging environment. The thread
amounts for IO and compute are now configurable through `--io-threads`
and `--compute-threads`.

NOTE: The 2 thread limit was reached with a very high RPS spike (~2000).
Changed the default to 4, also because buildernet instances have more
compute available.